Enumerate object files for both lib types
This commit will update the Makefile so that each object file gets compiled
properly for each build type (static and shared) on GNU/Linux.

In the future, it would be cool to set up an implicit pattern rule, but this had
some unsuccessful behavior in testing. Thus, this rule has been removed in this
commit.